backend/requirements.txt
Django==5.2.3
djangorestframework==3.16.0
django-cors-headers==4.7.0
python-decouple==3.8
mysqlclient==2.2.7
asgiref==3.8.1
sqlparse==0.5.3
tzdata==2025.2
Pillow==11.2.1 